refactor(skills-editor): extract constants to constants.js
